refactor: Modify package add lotto sub domain
- lotto 서브 도메인 추가 리펙토링
